pyproject.toml
[project]
name = "core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "langchain>=0.3.26",
    "langchain-core>=0.3.68",
    "markdown-it-py>=3.0.0",
    "mdit-py-plugins>=0.4.2",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "uvicorn>=0.35.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["ui", "core"]

[tool.uv.workspace]
members = [
    "oca_langchain",
]
